Added slide action.
Unfortunately, adding an entry in an enumeration list on the code side
isn't propagated to data referencing any enum in this list.
The data stores probably an integer value for the enum,
which gets out of sync if the enum value is changed due to some new enum
introduced in between.

So I went through players actions and action area prefabs to fix
everything.

Also a few addition to scene test_08 including a slide area.